pyproject.toml
[project]
name = "project-mailing"
version = "0.1.0"
description = ""
authors = [
    {name = "Maxim Andreev",email = "andrianov_maksim@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "redis (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
